Submission Creation/CS145 Project 3 Group 604480880 004651115 204411921/codes/requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.3
boto==2.38.0
chardet==2.3.0
configparser==3.5.0
crcmod==1.7
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
enum34==1.1.6
feather-format==0.4.0
functools32==3.2.3.post2
futures==3.2.0
google-compute-engine==2.7.5
gyp==0.1
html5lib==1.0.1
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.1.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.2.0
MarkupSafe==1.0
matplotlib==2.2.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.2
pandas==0.22.0
pandocfilters==1.4.2
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psutil==5.4.3
ptyprocess==0.5.2
pyarrow==0.8.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.9.1
scandir==1.7
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
terminado==0.8.1
testpath==0.3.1
tornado==5.0
tqdm==4.19.7
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.13.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
xgboost==0.7.post3
